+++ src/gtt_ghtml_deprecated.c
@@ -31,7 +31,7 @@
 #include "ghtml.h"
 #include "ghtml-deprecated.h"
 #include "proj.h"
-#include "util.h"
+#include "gtt_util.h"
 
 /* This file contains deprecated routines, which should go away
  * sometime in 2004 or 2005, around gnotime version 3.0 or so
