+++ src/gtt_idle_timer.h
