+++ src/gtt_log.c
@@ -25,7 +25,7 @@
 
 #include "cur-proj.h"
 #include "log.h"
-#include "prefs.h"
+#include "gtt_preferences.h"
 #include "proj.h"
 
 #define CAN_LOG ((config_logfile_name!=NULL)&&(config_logfile_use))
