+++ src/gtt_plug_in_edit.c
@@ -27,7 +27,7 @@
 #include "gconf-io.h"
 #include "menus.h"
 #include "plug-in.h"
-#include "util.h"
+#include "gtt_util.h"
 
 struct PluginEditorDialog_s
 {
