+++ src/gtt_queries.c
@@ -21,7 +21,7 @@
 #include <glib.h>
 #include <limits.h>
 
-#include "prefs.h"  /* XXX tmp hack for global config_daystart */
+#include "gtt_preferences.h"  /* XXX tmp hack for global config_daystart */
 #include "proj.h"
 #include "proj_p.h"
 #include "query.h"
