+++ src/gtt_util.h
